requirements.txt
aiocfscrape==1.0.0
beautifulsoup4==4.7.1
lxml==4.7.1
mypy==0.931
python-telegram-bot==13.10
PyYAML==6.0
types-PyYAML==6.0.3
